added types for events.ts
